+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/optimization/DataFlowVariableStates.java
@@ -462,7 +462,7 @@ class DataFlowVariableStates {
 
             // Do not report uninitialized reads in unreachable and dead states
             boolean report = reportUninitialized && !isolated && !dead && reachable && ixReachable;
-            if (report && !initialized.contains(variable) && variable.getType() != ArgumentType.BLOCK) {
+            if (report && !initialized.contains(variable) && variable.getType() != ArgumentType.BLOCK && variable.getType() != ArgumentType.BUILT_IN) {
                 print("!!! Detected uninitialized read of " + variable.toMlog());
                 optimizer.addUninitialized(variable);
             }